upgrade: Rework router migration to always use neutron-ha-tool
With this change we will always use neutron-ha-tool to evacuate router
off of a node that is going to be upgrade, regardless of wether this is
the first node to be upgraded or a subsequent one.

We make use of --wait-for-router to migrate router sequentially and wait
for the router migration to complete before going on with other work.

It also makes use of the "--ssh-delete-namespace" parameter that we need
when migrating routers from nodes which are not running the l3-agent
anymore.

The change also required some adjustments to the unit test of the drbd
upgrade case. The search stub for the drbd nodes needs to return two
nodes (master and slave) in order to work correctly now.