upgrade: enable ptf repo during the noderepochecks (bsc#1015827)
the ptf repo is not tracked under any feature so we need to enable
it by hand.

enabling it as a feature would not make sense as it may not be
available, and the api response would show it as missing and
make the repocheck appear as failed. Also conceptually the ptf
repo does not seem to be a feature.