barclamp: Make Config.load() more clever when no instance is specified
When no instance is specified and we can't detect which one to use, then
try with "default", and if that doesn't work, find the first instance
that has an applied proposal for this barclamp.

This makes it possible to do something like Config.load("openstack",
"nova") without having to care what's the instance of the applied nova
proposal -- which is okay because there's only one applied proposal for
that barclamp.