upgrade: Save the upgrade errors in consistent format
When the synchronous API call fails, it returns the error in a form
of 'data' and 'help' touple. We should save it the same way in the
progress file, especially because for aynchronous format, the content
of progress file is returned by the status API call.

(cherry picked from commit 44f7a55da5b79412480846d5de80603b3ab21559)