upgrade: Fix format of precheck errors in status
Before, the status API returned array of error hashes when some of
the prechecks failed. Everything was enclosed in one more hash.
This change converts this array to one hash and removes extra nesting
to match the common error format used in other places.
The error keys are assumed to be unique.