upgrade: Fix handling of multiple zypper prompts
There can be more than one <prompt> element in zypper xml output.
In such cases, Hash.from_xml() returns array of hashes instead of
a hash for those subtrees.
In this particular case, the simplest solution is to show only first
prompt as the user needs to complete the actions in sequence anyway
when running it in the terminal.