upgrade: Check if a feature/addon is deployed, not enabled
if we only check if it is enabled on the repository level, we will
have a false result within the repochecks in case a repository is
disabled, but necessary to have, because it has been deployed.

so if an ha repo becomes unavailable at some point during the upgrade
it will not even show up in the repochecks anymore, because it gets
disabled on the repository level once it is not available anymore.

(cherry picked from commit 1e5fbb083a075831c4ac3d774c29e62a72c40d78)