apache2: Do not define apache2 service twice (bsc#1021106)
This causes the magic to override the LWRP in crowbar-pacemaker to not
fully work, as it only overrides the second definition, and not the
first one. Which results in the :enable action to always be done for the
apache2 service, even when pacemaker is used.

https://bugzilla.suse.com/show_bug.cgi?id=1021106
(cherry picked from commit 1a1c30d4ee5965be654327f34bf656b9c1a71004)