upgrade: Error out on first failed router-migration (bsc#1020914)
This uses the newly introduced "--exit-on-first-failure" argument for
neutron-ha-tool to immediately exit when a single router migration
fails.

We want the user to give a chance to debug/fix failing router
migrations to prevent all routers going offline at once in case
something is wrong on the target nodes (e.g. non-running l2-agents).

https://bugzilla.suse.com/show_bug.cgi?id=1020914