upgrade: Enable preparing only ceph nodes for the upgrade
Independently on upgrading the Cloud, ceph nodes will need to be
upgraded as well. During the Cloud upgrade, we need to keep them
in crowbar_upgrade state so the chef-client is not running there.

So we provide extra option to execute the prepare_nodes_for_crowbar_upgrade
just for the ceph nodes, and this could be run from different part
of the code before the Cloud upgrade starts.

The same applies for canceling the upgrade: if we invoke
revert_nodes_from_crowbar_upgrade from the normal Cloud upgrade process,
we do not want to touch the state of ceph nodes which should be handled
independently.

(cherry picked from commit 03bec981d7ee6ab660544e1e2951c5c847fd899f)