provisioner: set blocksize option on tftp server to 1024
It seems that a variety of tftp clients can be picky about
the options, so there is a mechanism to restrict what
options the server offers to the clients to increase compatibility
with a wide variety of clients.

The qemu uefi tftp client doesnt seem to work very well negotiating the
blocksize. Forcing it to be 1024 makes it work.