upgrade: Extra page for preparing ceph nodes for the upgrade
Before the Cloud upgrade, ceph nodes must be manually upgraded to
the latest version of SES. Before that, we need to move them to
crowbar_upgrade state.
There is no UI for SES upgrade, but we can provide simple UI for
the preparation that is needed for both SES and following SOC upgrades.