deobfuscate the network conduit resolver
This is a forward-port of the first commit of #575 (893bfbb5) in an
attempt to make this code even clearer.  #575 never got merged because
it took too long to review it, and in the meantime all of this code
got significantly rewritten during extraction of the conduit resolver
code into a library to be shared between Chef and Crowbar.

The other (fix) part of #575 was superseded by #803.