network: Set MTU for VLAN parent interface (bsc#1024279)
If a physical interface has VLAN(s) configured with a custom MTU set but
the underlying physical network is not used directly and has no MTU in
its config, use the MTU from the VLAN interface for the parent device.
If there are multiple VLANs, use the highest MTU for the parent.

Also, if the physical parent from a VLAN interface is configured but the
MTU is lower than one of its VLAN interfaces, abort the chef-client run
and raise a useful error message.

(cherry picked from commit fc4152dd2aa903959e9f978b1fb48f51f7146420)