upgrade: Disable all current repos before starting the upgrade
As part of the node upgrade we're replacing all the product repositories
with the ones from the new product version. To be able to cleanup
correctly we need disable the old product repositories in crowbar (i.e.
remove the associated data bags) before the repository configuration
files (/etc/crowbar/repos.yml and ./crowbar_framework/config/repos-*.yml)
are replaced.

If the upgrade is canceled, we re-enable them.