upgrade: enhance status library to set a certain upgrade mode
This introduces a new attrbute "selected_upgrade_mode" to force a
certain upgrade mode ("normal" when "non-disruptive" would be possible.
It's stored separately from the mode computed by the prechecks to be able
to figure out if the user wants to override the prechecks.

This commit also brings back the "upgrade_mode" method to check which
upgrade mode is actually supposed to be used.