upgrade: Save more detailed information about current upgrade activity
It seems useful to show more about what we are doing to nodes,
especially when it takes more time.
This extends the previous information saved as part of current_node,
which is marked only at the beginning and end of upgrade.

(cherry picked from commit fd7fc1675e38baa5daed9471dadd5362bcb76811)