Try to fix tempfile handling for UpgradeStatus lib
The old way of creating the temporary filename for the status file
during the unit test caused a lot of test failures lately, because the
state file randomly disappeared during the test. This seemed to be
caused by our weird usage of Tempfile.open() and File.unlink().

This commit changes the code to generate a just temporary filename
(instead of actually creating and deleting the file again). File
deletion now happens explicitly in an "after"-hook.