upgrade: provide better help for pacemaker issues
Tries to provide a small help or refer to the HA docs in case of failures
on the cluster resources. Also moves the references of the node names
to the data field and pairs the node name with the error found at it.

(cherry picked from commit ba5f4bad3f53c556f788411471810fb37dc5f2c2)