Restore ovs bridge settings after reboot (bsc#1029179)
During a clean shutdown wicked will delete any ovs bridges that it feels
responsible for. OTOH, when the node is shutdown uncleanly (e.g. during
a fencing operation) the ovs bridge will not get deleted, which for the
bridges that are touched by the neutron-openvswitch-agent means they
will stay in "secure" mode (without any flows defined) even after
reboot.  This means that the node will be disconnected until
neutron-openvswitch-agent is started. To prevent this we now reset the
bridge's fail-mode before the interface is coming up.

This is a backport of 8229c1e2f2d26fe2102ee935f5104e9f9fe3dac5 from
master, with the pieces removed that do not apply to stable/3.0