Restore ovs bridge settings after reboot (bsc#1029180, bsc#1029179)
During a clean shutdown wicked will delete any ovs bridges that it feels
responsible for. This means that any customer settings on tha brigde
will get lost and need to be restored. As we set a custom datapath_id
for our bridge to prevent bsc#1022074. We need to restore that setting
when the bridge is recreated during startup. Normally chef-client would
do that as part of crowbar_join, but we also need it to happen when the
admin server is down for some reason.

OTOH, when the node is shutdown uncleanly (e.g. during a fencing
operation) the ovs bridge will not get deleted, which for the bridges
that are touch by the neutron-openvswitch-agent means they will stay in
"secure" mode (without any flows defined) even after reboot. This means
that the node will be disconnected until neutron-openvswitch-agent is
started. To prevent this we now reset the bridge's fail-mode before the
interface is coming up.