upgrade: Do not try to disable non-existing service
At SP2, systemctl disable complains when disabling service that is not installed.
This is normally not a problem, since we execute this code at SP1 (Cloud6)
nodes only, but it can fail in the corner case when user has Ceph nodes, upgrades them
(so they have SP2 and are staying in the crowbar-upgrade state) and the redeployes
some barclamp (e.g. for fixing some setting required for the Cloud upgrade).