Unclaimed disks: improve handling of multipath devices (bsc#1031065)
dm_multipath devices were being skipped while its paths were being included
in the array of unclaimed devices, allowing the DRBD recipe to claim them
and later fail while running pvcreate.

Introduce two new methods in Disk for checking whether the device (or its
holder) is named as "mpath-*".

This will make sure that when having a multipath device with several holders
we add to the unclaimed list only the multipath device and we ignore
the holders.