upgrade: Move the freshly upgraded remote node to ready state
After remote compute node is upgraded, nova-compute service is not
running there, because the services are managed by the corosync cluster,
It would have been started with next chef-client run on controller node,
but we cannot wait for that.
So let's speed things up by manually proclaiming the node as ready.