Fix node bulk edit to save only the changed nodes (bsc#1008838)
We were not excluding the admin node from the target_platform and license_key checks as
it doesnt have an allocated value sent on POST.
Also, we didnt check for a blank value (nil or empty or "") in the public_name values.
Both of this issues triggered a save of all nodes, even if the values were the same.