bsc#1033537: Fix domain caching
The Node model goes through Crowbar::Settings to get the domain name of
a node. Without this patch, Crowbar::Settings will use the domain
attribute of the dns barclamp and, to avoid repeated database queries,
store it permanently as a class variable. This is a problem because if
the domain attribute is changed in the dns barclamp, the change will not
be reflected in the Node model, which means that if a node is removed
and re-added, its new domain name will be properly reflected in Chef but
crowbar will respond with an HTTP 404 Not Found when a node is queried.

This "caching" strategy has been part of the code apparently since the
beginning of time[1] and it's unclear what the performance gain was
thought to be at the time, so one might think that a reasonable option
to fix this would be to undo the "caching" strategy and simply query the
Proposal relation when the domain needs to be fetched. However, this is
not reasonable because the crowbar dashboard UI queries the node list
every few seconds and this causes a lookup of the domain for each node,
so with a reasonably large number of nodes this would easily overload
the database server. Rails' own SQL caching mechanism is not sufficient
here because query caches are destroyed at the end of an action[2].

This patch solves the problem by writing the proposal revision number
to Rails' cache when the dns barclamp is updated, and storing both the
revision number and the domain attribute as a class variable in the
Crowbar::Settings class. The Crowbar::Settings class then queries the
Rails cache to see if the dns proposal has been updated, and if so
updates its copy of the domain attribute. This way we avoid repeated
trips to the database but changes to the proposal will be propogated.

[1] https://github.com/crowbar/crowbar/commit/f95325fc7ec27a2c8ed4a184eca95af8f428ac3d
[2] http://guides.rubyonrails.org/caching_with_rails.html#sql-caching