upgrade: Make sure to upgrade all nodes in normal node (bsc#1035528)
It seems that (probably because of some manual interventions) there
is a chance that node has a role asociated in the proposal, but
the role is not known for node itself.

So with this change, we make sure all nodes are upgraded. If they
cannot be identified as compute nodes by searching for nova-compute role,
they are upgraded as 'remaning' (non-compute and non-controller) ones.