crowbar: Do not save applied proposal as role too early in apply_role
Until now, we saved the role for the applied proposal early in
apply_role. However, apply_role could fail after that but still before
apply_role_pre_chef_client, resulting in the role containing incomplete
data (since apply_role_pre_chef_client is used to process the role and
add some info, like "use HA").

In case of such a failure (and that can happen when we fail to get a
lock for a node), then a periodic chef-client run would use a role with
incomplete information, possibly leading to changing the config on the
node wrongly.

This partially addresses
https://bugzilla.suse.com/show_bug.cgi?id=857375

Closes https://github.com/sap-oc/crowbar-openstack/issues/34