dns: Disable domain field in UI (bsc#1033537)
The domain field should not be changed after the initial deployment.
Chef uses this field to identify nodes, including the admin node, and
changing it causes crowbar and chef to be able to unable to fetch data
about a node. This changes makes the field in the DNS barclamp UI
disabled and adds a note about it to discourage trying to make this
change.