network: set max mtu of parent to max of children
iterate over all the children to obtain the max mtu that we need to
set to a parent nic. Before this, only the first vlan children of a parent
was used to set an mtu value, failing afterwards if the next vlan had a bigger
value.

(cherry picked from commit da1f384f25299b073cfe5974262dfe4e9f94c8dd)