network: Add support for xmit_hash_policy (bsc#1033917)
Add support for the xmit_hash_policy property of NIC bonding to be queried
and set, as well as adding the existing miimon property and new
xmit_hash_policy property to the network schema.

Add support for miimon to be set (or overloaded) on a per-network basis,
like xmit_hash_policy.

Drive-by correctly setting miimon rather than hardcoding 100 for the
interfaces(5) template.