network: set max mtu of a parent to the max of its children
iterate over all the children to obtain the max mtu that we need to
set to a parent nic. Before this, only the first vlan children of a parent
was used to set an mtu value, failing afterwards if the next vlan had a bigger
value.

(cherry picked from commit c108dba2b9556840785411b2c4ab7d15c8bdda86)