provisioner: Use discovery mode as default for writing node boot files
Right now, we were defining the content of the boot files for a node
without any default (so it was either "install files" if installing, or
"discovery files" in other cases).

However, even when installing, we may want to go back to discovery mode
to re-trigger some code that may have failed (like admin IP address
allocation if the range was temporary full).

So move some code around for that.