Enhance network-json-resolve to also work with exported json for nodes
When debugging data from a deployment that we don't have access to, we
usually only get json describing the nodes and the network.json file.
Without direct access to the chef-server, we can't query chef to get the
data about the node.

So add a --node-json option to allow not using chef to get that
information. If used together with --network-json, it means we need no
access to chef-server at all.