crowbar: Fix re-running chef after failure in apply_role for admin nodes
When a chef call fails in apply_role, we try a second time to run chef.
However, for admin nodes, we were not remembering the pid/host pair
correctly, resulting in the second chef call to be invalid (for no
host).

This resulted in the creation of /var/log/crowbar/chef-client/.log,
which is a long-standing bug.