crowbar: wait for all chef clients at the same time
Before this we moved the wait for the chef clients to
each thread. This could potentially mean that we would have several
threads locked awaiting for their nodes chef client to finish, and
the thread pool dedicated to locking+waiting could decrease as we
encounter more and more chef runs.

Instead of that, lets lock everything on threads as soon as possible so
no new chef clients can be run and then wait for all the nodes to finish
their chef runs.