crowbar: Fix the wait for the chef daemons
Before the threading patch we were waiting for ALL of the nodes
being applied to finish their chef runs. After what looked like a typo
we were only looking at the nodes locked, which ignores the admin node.
So it could happen that the admin node was running the chef client
and it was ignored, not filling properly the templates for bind
thus not allowing the other dns server nodes to transfer zones,
which made the multi-dns test fail all the time randomly

(cherry picked from commit af14a7b70248e7890274282dc6816f48a27bba98)