Fix network_validator (bsc#1025642)
The previous network validator take the first private IP from the
system and expect that this is the admin IP.  This work OK in case
that eth0 is listed before eth1, but this is something that do not
happends always.

This change use the FQDN information to deduce the admin IP, as
install-suse-cloud is doing.