test: enhance the test-kitchen tests
Updated the tests to be more extensive and created a profile instead of
separated tests as an example so we can bundle libraries to use as
resources with inspec.

Add a zypper_package custom resource as example on how to create
custom resources for testing.

Expand the default opensuse image with net-tools so we can use netstat
to check for running ports