crowbar: Don't store invalid locks (bsc#1055669)
Currently, if a locking failure occurs, the lock variable will be unset
and an empty lock will be added to the lock queue. This is a problem
when it comes time to release locks, where the error "undefined method
`release' for nil:NilClass" will occur and cause a proposal failure.
This patch ensures that only successfully acquired locks make it into
the lock queue.