Retry to resolve a hostname before continue
Before running commands we check if a node is ready or the chef-client
is still running. Before doing that, we try to resolve the hostname
and continue with the resolved IP.
It seems that in some cases, resolve_host() does not work and returns
an empty string. The crowbar production.log then tells us:

INFO -- Waiting for already running chef-client on
  d52-54-77-77-01-03.vk2.cloud.suse.de (); next attempt in 10 seconds
  until 2017-08-31 04:36:53 +0000

but inside of "()" should be the resolved IP address. This message is
repeated until we reach the timeout. This repetition is useless
because we retry with an empty IP.
To minimize the impact of this problem, retry to resolve the hostname
for some time.