utils: override the service provider to allow for no-restart of services
We override the default service provider from chef to allow checking if
restarts are diallowed. The flags for restarts disallowed is stored on
the databag crowbar-config -> disallow_restart and consists of the
cookbook name and if disallow restarts is true/false.

This allows us to provide an attribute on the barclamp level that will
skip the service restart if the end user chooses to.

Service restart operations will also maintain a data structure so that
if a service restart was skipped, that restart request will be
reigstered. The data is stored in:

node[:crowbar_wall][:requires_restart]

A hash is stored under the above path. Keys are service names and values
are hashes with two additional keys. An example:

{
  service_name => {
    pacemaker_service: is_pacemaker_service,
    timestamp: Time.now.getutc
  }
}

Whenever a restart actually happens, the restart request for the
specific cookbook/service is cleared.

For apache2 service restarts, as we are unable to know which cookbook
trigerred it, we refer to a temporal key set in the origin cookbooks

node.run_state["apache2_restart_origin"]

that let us know which service triggered the apache2 restart.

(cherry picked from commit 5f6171f9dd7f133cb646e09a912fb9b5135886e9)