ipmi: Option to disable BMC NAT
NATing IPMI access through Admin Server is not ideal in some cases.
STONITH based on this can trigger fencing if Admin Server is (temporarily)
not accessible.
This change adds option to disable BMC NAT feature.
Access from nodes to IPMI plane needs to be configured externally for this
to work properly.