utils: Add utils_systemd_service_restart LWRP
This resource adds (or removes) a systemd config snippet for a service,
to change the behavior to Restart=on-failure for that service. This
means systemd will automatically restart the service on failures, which
makes a relatively good self-healing solution without pacemaker.

See systemd.service(5) for details on Restart=

Note that the fetch_name_from_service_resource attribute of this
resource enables automatically fetching the systemd service name from
the service resource, which helps remove duplication of code in the
cookbooks when the service name is being overridden in the attributes of
the service resource. For instance, this will work as expected:

  service "foobar" do
    service_name "notfoobar"
  end

  utils_systemd_service_restart "foobar"

(cherry picked from commit 030f1e4d1e8d619fe34e37b94a73bb4f2e25435f)