dns: fail on empty server list
to not screw up existing resolv.conf and dnsmasq.conf
which would make it hard to leave this state

Why is this change necessary?
Discovered when working on http://bugzilla.novell.com/show_bug.cgi?id=1058017
Nodes that are not 'ready' are not considered available to serve as DNS server,
but when cloud-operators don't keep the dns-server role on the admin-node,
this can lead to situations were nodes drop all entries from /etc/resolv.conf and dnsmasq.conf and cannot resolve any DNS anymore
which could make it hard to leave this state,
because rabbitmq and possibly others (especially with SSL) need DNS to work.

How does it address the issue?
we detect that case and raise a (fatal) exception
to not screw up existing resolv.conf and dnsmasq.conf

Is there additional information worth sharing?
There is probably more to that bug, such as nodes losing roles, not just config flags under certain conditions.
There will still be manual intervention necessary to get DNS-servers back.