all: Make systemd restart services on failures
This offers some automatic recovery, which is nice. It turns out it's
also recommended by systemd documentation, in case a service is
long-running, which is the case here.

It's worth mentioning that in the past, the periodic chef-client has
been providing this, to some extent. But this means the recovery could
take up to 15 minutes. It was actually not working in the case of the
chef server infrastructure (chef-server, chef-solr, couchdb, etc.), and
systemd can really help here.

This is of course disabled when a service is managed by Pacemaker, as we
do not want to have systemd interfere with Pacemaker in that case. Note
that this commit doesn't touch any service managed by Pacemaker, though.