crowbar: Add skip_unready_nodes experimental option
Adds to the experimental options the skip_unready_nodes
configs that would allow to set a list of roles for which we would not
have to wait for all nodes to be in status ready to apply the barclamps.

Removed experimental option that not exists in this version.
Added: Load experimental configuration file

(cherry picked from commit dd4e0ef3a764a19d5d0f1db4c8459907cc21f4de)