crowbar: remove unready nodes from deployment
If we set the experimental skip_unready_nodes config to true and provide
a list of roles, we try to skip nodes that are not ready from the
deployment of the barclamp so we dont need all nodes to be ready to
deploy a barclamp.

This is useful when dealing with a high number of nodes (i.e.
nova-compute) and we dont need all those other nodes to be ready to add
a new node to a role

(cherry picked from commit e8a59469e7e1ccc252c376d2dce9a3c1d948ddfa)