crowbar: Fill pre_cached_nodes with one search
As we reset the pre_cached_nodes before calling
remote_chef_client_threads the caceh will be empty and the nodes will
have to be loaded one by one, which in a higher number of nodes
translates to a higher number of chef api calls.

Instead, pre-fill the pre_cached_nodes by doing batch searches of 50
nodes. This should help with performance when we have a high number
of nodes to run chef onto.