crowbar_framework: Add the RestartManagementController
Adds the base controller, routes, tests, endpoints for the
RestartManagementController.

This allows and end user to see at a glance all the *active* openstack
barclamps and the policy for the service reboot (enabled/disabled service reboots)

It also allows to see if any services require a manual restart and clear
those flags to signal a manual restart of a service.

Adds 2 new helpers to Crowbar::DataBagConfig to help with databag
manipulation, to deal with non-existant databags or items