crowbar: Forward protocol to rails (bsc#1059733)
Without this patch, when crowbar is deployed with the HTTP+HTTPS option
and the user is connected with HTTPS, any redirect (such as after a
proposal is applied or when a node is deleted) will force the user back
to an HTTP connection. This is obviously insecure. This patch adds a
setting to the crowbar apache vhost to forward the request protocol to
rails so that the ssl property of the rails request object will be set
and the redirect_to method in rails will do the right thing.

(cherry picked from commit 7970e517ec619f254b6276c81213d29b81977728)