s390x: recognize DASD drives
DASD drives in s390x use a different name schema than other
devices, like sda or hda. Change the regular expression to
recognize them.

Indirectly this fix a bug in crowbar, where the nodes appears
without hard disk information. Also is required for the Cinder
barclamp, in order to create the partition and assign it to the
cinder-volume group.