upgrade: Don't fail without openstack db (bsc#1061777)
If only the admin (crowbar) node is deployed, an upgrade from SOC6 to
SOC7 currently fails with:

Creating OpenStack database backup failed
undefined method `[]' for nil:NilClass

There is no node with the role "database-config-default" in that
case. Do not fail but continue the normal upgrade process if the role
was not found.