utils: Fix restart flag removal
It looks like when you use fetch to fetch an attribute from the node
object we get the direct value, which if modified (deleted/updated)
doesnt get to be reflected into the node object, so when we deleted the
restart flag from a service and the node was saved, the modification we
did was not saved.

Instead, lets just ask for forgiveness when trying to delete the
attribute by using rescue to save from any issues of attributes not
existing when deleting the flags.

(cherry picked from commit 9d19923bf4dc4a3c75dcb16d4cfcd2ec4fdbbb63)