ntp: Make listening net configurable (bsc#1047941)
In cases where the admin network is firewalled from the outside world,
the cloud ntp servers will not be able to connect to external ntp
servers to synchronize. This patch adds the option to set additional
listen interfaces for nodes with the ntp-server role so those ntp
servers can sync with outside servers.

On the ntp-server host, only packets from configured external NTP servers
and from hosts in the admin network are allowed; everything else is
ignored. In addition to that, the rate-limiting feature has been
enabled for requests coming from the external NTP servers.