ohai: Collect permanent address of NICs
When network interfaces are using bonds, they actually all start using
the same MAC address, and their real MAC address (called permanent
address) is hidden. This is information that we didn't collect in ohai,
and we will need to ensure we correctly configure the DHCP server for
all permanent MAC addresses, to ensure PXE boot is always doing the
right thing.