provisioner: Use also permanent addresses for dhcp
When using bonds, all slave interfaces adopt the same mac address, and
therefore their permanent mac address is "hidden". This results in the
DHCP config not being fully complete since some MAC addresses are not
known, and therefore we cannot configure anything for them.

This may lead to nodes rebooting and reinstalling because there's no
DHCP config telling the node to not boot from that interface, or to boot
from disk, and therefore this falls back to the discovery image, which
will trigger some automatic reinstall (since the node is known).